test: only test on supported node versions
BREAKING CHANGE:
Tells TravisCI to run tests on supported NodeJS versions. The full
release and support schedule can be found on the NodeJS website:
https://github.com/nodejs/Release#release-schedule.  Node 6.x is still
in maintenance LTS until April 2019.

By dropping unsupported versions, we are able to keep pace with the
css-nano library newer versions and addresses the errors in #104.